+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Rename big-ruby-program.rb to big-ruby-program
 
 - chapter-3.rb [examples from chapter-2]
+- chapter-4.rb [examples from chapter-4]
 
 ## [0.0.1] (May 8 2021)